+++ lib/textus/cli/verb/audit.rb
@@ -9,6 +9,7 @@ module Textus
         option :role_filter, "--role=ROLE"
         option :verb_filter, "--verb=V"
         option :since, "--since=ISO8601|RELATIVE"
+        option :seq_since, "--seq-since=N"
         option :correlation_id, "--correlation-id=ID"
         option :limit, "--limit=N"
 
@@ -21,6 +22,7 @@ module Textus
             role: role_filter,
             verb: verb_filter,
             since: since_time,
+            seq_since: seq_since&.to_i,
             correlation_id: correlation_id,
             limit: limit&.to_i,
           )